+++ src/Gel.cpp
@@ -7,7 +7,7 @@
 //
 
 #include "Gel.h"
-#include "GelParticleSource.h"
+#include "Sample.h"
 
 using namespace std;
 using namespace ci;
@@ -18,87 +18,102 @@ const vec2 kPosVec (0,1);
 void Gel::setLayout(
 	float		lane_dimension, 
 	float		pos_elec_dimension,
-	int			numLanes )
+	int			numLanes,
+	float		ymargin )
 {
+	mYMargin = ymargin;
+	
 	mSize = vec2( lane_dimension, pos_elec_dimension );
 	
 	mNumLanes  = numLanes;
 	mLaneWidth = mSize.x / (float)numLanes; 
 }
 
-void Gel::insertSamples( const GelParticleSource& src, int lane, int num )
+void Gel::insertSample( const Sample& src, int lane )
 {
-	vec2 laneLoc = kLaneVec * ((float)lane*mLaneWidth + mLaneWidth/2.f);
+	vec2 laneLoc = vec2(0.f,mYMargin) + kLaneVec * ((float)lane*mLaneWidth + mLaneWidth/2.f);
 	
 	vec2 size;
 	size.x = mLaneWidth * .25f;
-//	size.y = mLaneWidth * .05f;
-	size.y = (mSize.y) / (float)num; // try to dynamically scale based on # samples...
+	size.y = mLaneWidth * .05f;
 	
-	for( int i = 0; i<num; ++i )
+	for( auto frag : src.mFragments )
 	{
-		GelParticleSource::Result gen = src.next(mRand);
+		Band b;
+
+		b.mBases		= frag.mBases;
+		b.mMass			= frag.mMass;
+		b.mDegrade		= frag.mDegrade;
+				
+		b.mStartLoc		= b.mLoc = laneLoc;
 		
-		Particle p;
+		float a = min( 1.f, b.mMass / 125.f ); // alpha set to mass 125ml
 		
-		p.mStartLoc		= p.mLoc = laneLoc;
-		p.mSpeed		= gen.mSpeed;
-		p.mCreateTime	= mTime;
-		p.mExists		= true;
-		p.mColor		= ColorA(1.f,1.f,1.f,.1f); // just make up a color for now...
-		p.mSize			= size;
+		b.mCreateTime	= 0.f; // always start it at the start (not mTime); otherwise is kind of silly...
+		b.mExists		= true;
+		b.mColor		= ColorA(1.f,1.f,1.f,a);
+		b.mSize			= size;
 		
-		mParticles.push_back(p);
+		mBands.push_back(b);
 	}
 	
-	mDuration = calcDuration();
+	updateBandsWithTime(mTime);
 }
 
 void Gel::clearSamples()
 {
-	mParticles.clear();
+	mBands.clear();
 }
 
 void Gel::stepTime ( float dt )
 {
 	mTime += dt;
 	mTime = min( mTime, mDuration );
-	updateParticlesWithTime(mTime);
+	updateBandsWithTime(mTime);
 }
 
 void Gel::setTime( float t )
 {
 	mTime = t;
-	updateParticlesWithTime(mTime);
+	updateBandsWithTime(mTime);
 }
 
-float Gel::calcDuration() const
+void Gel::updateBandsWithTime( float t )
 {
-	if ( mParticles.empty() ) return 0.f;
-	
-	float d = MAXFLOAT;
-	
-	for( auto &p : mParticles )
+	for( auto &b : mBands )
 	{
-		// assume created at the top for now to simplify this...
-		d = min( d, mSize.y / p.mSpeed );
+		b.mExists = t >= b.mCreateTime;
+		
+		float bandTime = max(t - b.mCreateTime,0.f);
+		
+		float y = getYForBases( b.mBases, bandTime );
+		
+		b.mLoc = b.mStartLoc + kPosVec * y;
 	}
-	
-	return d;
 }
 
-void Gel::updateParticlesWithTime( float t )
+float Gel::getYForBases( int bases, float t ) const
 {
-	for( auto &p : mParticles )
-	{
-		p.mExists = t >= p.mCreateTime;
+	// map domain to 0..1
+	float y = (float)bases / 10000.f;
+		// y=1 means lots of bases, slow, top of gel
+		// y=0 means few bases, fast, bottom of gel
+
+	// time
+	y = 1.f - t * (1.f - y);
 		
-		float particleTime = max(t - p.mCreateTime,0.f);
-		
-		float travel = p.mSpeed * particleTime;
-		
-		travel = min( travel, mSize.y ); // no falling off the edge
-		
-		p.mLoc = p.mStartLoc + kPosVec * travel;
-	}
+	// output is non-linear
+	const float K = .5f;
+//	float K = lerp( 0.f, 2.f, t );
+	
+	y = powf( y, K );
+	
+	// flip it, so that y=1 puts us at bottom of gel
+	y = 1.f - y; 
+	
+	// scale
+	y *= mSize.y - mYMargin*2.f;
+	
+	// done
+	return y;
 }