+++ Changes.md
@@ -18,7 +18,7 @@
 - Can drop images into window, move them around 
 - Refactor gel view into new View system; simplify Gel layout params
 - Can drag around Gel and Timeline is attached. Aesthetic tweaks to Timeline.
-- Can drop in GelParticleSource .xml file, which makes a view object for it!
+- Can drop in Sample .xml file, which makes a view object for it!
 - Next:
 	- Allow pouring into gel lanes.
 	- Icon for particle source (cute beaker!); in multiple colors (specified in xml)?